chore(query): remove udf parser in ast crate
