+++ src/common/legacy-parser/Cargo.toml
@@ -9,11 +9,9 @@ doctest = false
 test = false
 
 [dependencies]
-common-ast = { path = "../../query/ast" }
 common-datavalues = { path = "../datavalues" }
 common-exception = { path = "../exception" }
 common-functions = { path = "../functions" }
-common-meta-types = { path = "../../meta/types" }
 common-planners = { path = "../planners" }
 
 async-trait = "0.1.56"
