+++ src/common/legacy-parser/src/analyzer/analyzer_expr_sync.rs
@@ -13,27 +13,25 @@
 // limitations under the License.
 
 use async_trait::async_trait;
-use common_ast::udfs::UDFDefinition;
-use common_ast::udfs::UDFExprTraverser;
-use common_ast::udfs::UDFExprVisitor;
-use common_ast::udfs::UDFFetcher;
-use common_ast::udfs::UDFParser;
-use common_ast::udfs::UDFTransformer;
 use common_datavalues::prelude::*;
 use common_datavalues::type_coercion::merge_types;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_functions::aggregates::AggregateFunctionFactory;
-use common_functions::is_builtin_function;
-use common_meta_types::UserDefinedFunction;
 use common_planners::Expression;
+use sqlparser::ast::BinaryOperator;
+use sqlparser::ast::DataType as AstDataType;
 use sqlparser::ast::DateTimeField;
 use sqlparser::ast::Expr;
+use sqlparser::ast::Function;
+use sqlparser::ast::FunctionArg;
 use sqlparser::ast::FunctionArgExpr;
 use sqlparser::ast::Ident;
 use sqlparser::ast::Query;
+use sqlparser::ast::TrimWhereField;
 use sqlparser::ast::UnaryOperator;
 use sqlparser::ast::Value;
+use sqlparser::ast::WindowSpec;
 
 use crate::analyzer_value_expr::ValueExprAnalyzer;
 use crate::sql_common::SQLCommon;
@@ -51,7 +49,7 @@ impl ExpressionSyncAnalyzer {
         let mut stack = Vec::new();
 
         // Build RPN for expr. Because async function unsupported recursion
-        for rpn_item in &ExprRPNBuilder::build(expr, vec![])? {
+        for rpn_item in &ExprRPNBuilder::build(expr)? {
             match rpn_item {
                 ExprRPNItem::Value(v) => Self::analyze_value(v, &mut stack, SQLDialect::MySQL)?,
                 ExprRPNItem::Identifier(v) => self.analyze_identifier(v, &mut stack)?,
@@ -342,24 +340,27 @@ impl ExpressionSyncAnalyzer {
     }
 }
 
-enum OperatorKind {
+pub enum OperatorKind {
     Unary,
     Binary,
     Other,
 }
 
-struct FunctionExprInfo {
-    name: String,
-    args_count: usize,
-    kind: OperatorKind,
+pub struct FunctionExprInfo {
+    pub name: String,
+    pub distinct: bool,
+    pub args_count: usize,
+    pub kind: OperatorKind,
+    pub parameters: Vec<Value>,
+    pub over: Option<WindowSpec>,
 }
 
-struct InListInfo {
-    list_size: usize,
-    negated: bool,
+pub struct InListInfo {
+    pub list_size: usize,
+    pub negated: bool,
 }
 
-enum ExprRPNItem {
+pub enum ExprRPNItem {
     Value(Value),
     Identifier(Ident),
     QualifiedIdentifier(Vec<Ident>),
@@ -378,40 +379,45 @@ impl ExprRPNItem {
     pub fn function(name: String, args_count: usize) -> ExprRPNItem {
         ExprRPNItem::Function(FunctionExprInfo {
             name,
+            distinct: false,
             args_count,
             kind: OperatorKind::Other,
+            parameters: Vec::new(),
+            over: None,
         })
     }
 
     pub fn binary_operator(name: String) -> ExprRPNItem {
         ExprRPNItem::Function(FunctionExprInfo {
             name,
+            distinct: false,
             args_count: 2,
             kind: OperatorKind::Binary,
+            parameters: Vec::new(),
+            over: None,
         })
     }
 
     pub fn unary_operator(name: String) -> ExprRPNItem {
         ExprRPNItem::Function(FunctionExprInfo {
             name,
+            distinct: false,
             args_count: 1,
             kind: OperatorKind::Unary,
+            parameters: Vec::new(),
+            over: None,
         })
     }
 }
 
-struct ExprRPNBuilder {
+pub struct ExprRPNBuilder {
     rpn: Vec<ExprRPNItem>,
-    udfs: Vec<UserDefinedFunction>,
 }
 
 impl ExprRPNBuilder {
-    pub fn build(expr: &Expr, udfs: Vec<UserDefinedFunction>) -> Result<Vec<ExprRPNItem>> {
-        let mut builder = ExprRPNBuilder {
-            rpn: Vec::new(),
-            udfs,
-        };
-        UDFExprTraverser::accept(expr, &mut builder)?;
+    pub fn build(expr: &Expr) -> Result<Vec<ExprRPNItem>> {
+        let mut builder = ExprRPNBuilder { rpn: Vec::new() };
+        ExprTraverser::accept(expr, &mut builder)?;
         Ok(builder.rpn)
     }
 
@@ -458,7 +464,10 @@ impl ExprRPNBuilder {
                 self.rpn.push(ExprRPNItem::Function(FunctionExprInfo {
                     name: function.name.to_string(),
                     args_count: function.args.len(),
+                    distinct: function.distinct,
                     kind: OperatorKind::Other,
+                    parameters: function.params.to_owned(),
+                    over: function.over.clone(),
                 }));
             }
             Expr::Cast {
@@ -570,29 +579,251 @@ impl ExprRPNBuilder {
     }
 }
 
-#[async_trait]
-impl UDFFetcher for ExprRPNBuilder {
-    fn get_udf_definition(&self, name: &str) -> Result<UDFDefinition> {
-        let udf = self.udfs.iter().find(|udf| udf.name == name);
+pub struct ExprTraverser;
 
-        if let Some(udf) = udf {
-            let mut udf_parser = UDFParser::default();
-            let definition = udf_parser.parse(&udf.name, &udf.parameters, &udf.definition)?;
-            return Ok(UDFDefinition::new(udf.parameters.clone(), definition));
-        }
-        Err(ErrorCode::UnknownUDF(format!("Unknown Function {}", name)))
+impl ExprTraverser {
+    pub fn accept<V: ExprVisitor>(expr: &Expr, visitor: &mut V) -> Result<()> {
+        let expr = visitor.pre_visit(expr)?;
+        visitor.visit(&expr)?;
+        visitor.post_visit(&expr)
     }
 }
 
 #[async_trait]
-impl UDFExprVisitor for ExprRPNBuilder {
+pub trait ExprVisitor: Sized + Send {
     fn pre_visit(&mut self, expr: &Expr) -> Result<Expr> {
-        if let Expr::Function(function) = expr {
-            if !is_builtin_function(&function.name.to_string()) {
-                return UDFTransformer::transform_function(function, self);
+        Ok(expr.clone())
+    }
+
+    fn visit(&mut self, expr: &Expr) -> Result<()> {
+        match expr {
+            Expr::Nested(expr) => ExprTraverser::accept(expr, self),
+            Expr::Value(value) => self.visit_value(value),
+            Expr::Identifier(ident) => self.visit_identifier(ident),
+            Expr::CompoundIdentifier(idents) => self.visit_identifiers(idents),
+            Expr::IsNull(expr) => self.visit_simple_function(expr, "is_null"),
+            Expr::IsNotNull(expr) => self.visit_simple_function(expr, "is_not_null"),
+            Expr::UnaryOp { op, expr } => self.visit_unary_expr(op, expr),
+            Expr::BinaryOp { left, op, right } => self.visit_binary_expr(left, op, right),
+            Expr::Exists(subquery) => self.visit_exists(subquery),
+            Expr::Subquery(subquery) => self.visit_subquery(subquery),
+            Expr::Function(function) => self.visit_function(function),
+            Expr::TryCast { expr, data_type } => self.visit_try_cast(expr, data_type),
+            Expr::Cast {
+                expr,
+                data_type,
+                pg_style,
+            } => self.visit_cast(expr, data_type, pg_style),
+            Expr::TypedString { data_type, value } => self.visit_typed_string(data_type, value),
+            Expr::Position {
+                substr_expr,
+                str_expr,
+            } => self.visit_position(substr_expr, str_expr),
+            Expr::Substring {
+                expr,
+                substring_from,
+                substring_for,
+            } => self.visit_substring(expr, substring_from, substring_for),
+            Expr::Between {
+                expr,
+                negated,
+                low,
+                high,
+            } => self.visit_between(expr, negated, low, high),
+            Expr::Tuple(exprs) => self.visit_tuple(exprs),
+            Expr::InList { expr, list, .. } => self.visit_inlist(expr, list),
+            Expr::Extract { field, expr } => self.visit_extract(field, expr),
+            Expr::MapAccess { column, keys } => self.visit_map_access(column, keys),
+            Expr::Trim { expr, trim_where } => self.visit_trim(expr, trim_where),
+            Expr::Array(exprs) => self.visit_array(exprs),
+            other => Result::Err(ErrorCode::SyntaxException(format!(
+                "Unsupported expression: {}, type: {:?}",
+                expr, other
+            ))),
+        }
+    }
+
+    fn post_visit(&mut self, _expr: &Expr) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_inlist(&mut self, expr: &Expr, list: &[Expr]) -> Result<()> {
+        ExprTraverser::accept(expr, self)?;
+        for expr in list {
+            ExprTraverser::accept(expr, self)?;
+        }
+        Ok(())
+    }
+
+    fn visit_tuple(&mut self, exprs: &[Expr]) -> Result<()> {
+        match exprs.len() {
+            0 => Err(ErrorCode::SyntaxException(
+                "Tuple must have at least one element.",
+            )),
+            1 => ExprTraverser::accept(&exprs[0], self),
+            _ => {
+                for expr in exprs {
+                    ExprTraverser::accept(expr, self)?;
+                }
+
+                Ok(())
+            }
+        }
+    }
+
+    fn visit_wildcard(&mut self) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_value(&mut self, _value: &Value) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_identifier(&mut self, _ident: &Ident) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_identifiers(&mut self, _idents: &[Ident]) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_exists(&mut self, _subquery: &Query) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_subquery(&mut self, _subquery: &Query) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_function_arg(&mut self, arg_expr: &FunctionArgExpr) -> Result<()> {
+        match arg_expr {
+            FunctionArgExpr::Expr(expr) => ExprTraverser::accept(expr, self),
+            FunctionArgExpr::Wildcard => self.visit_wildcard(),
+            FunctionArgExpr::QualifiedWildcard(_) => Err(ErrorCode::SyntaxException(std::format!(
+                "Unsupported QualifiedWildcard: {}",
+                arg_expr
+            ))),
+        }
+    }
+
+    fn visit_function(&mut self, function: &Function) -> Result<()> {
+        for function_arg in &function.args {
+            match function_arg {
+                FunctionArg::Named { arg, .. } => self.visit_function_arg(arg)?,
+                FunctionArg::Unnamed(arg) => self.visit_function_arg(arg)?,
+            };
+        }
+
+        if let Some(over) = &function.over {
+            for partition_by in &over.partition_by {
+                ExprTraverser::accept(partition_by, self)?;
+            }
+            for order_by in &over.order_by {
+                ExprTraverser::accept(&order_by.expr, self)?;
             }
         }
 
+        Ok(())
+    }
+
+    fn visit_cast(
+        &mut self,
+        expr: &Expr,
+        _data_type: &AstDataType,
+        _pg_style: &bool,
+    ) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_try_cast(&mut self, expr: &Expr, _data_type: &AstDataType) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_typed_string(&mut self, _data_type: &AstDataType, _value: &str) -> Result<()> {
+        Ok(())
+    }
+
+    fn visit_simple_function(&mut self, expr: &Expr, _name: &str) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_unary_expr(&mut self, _op: &UnaryOperator, expr: &Expr) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_binary_expr(&mut self, left: &Expr, _op: &BinaryOperator, right: &Expr) -> Result<()> {
+        ExprTraverser::accept(left, self)?;
+        ExprTraverser::accept(right, self)
+    }
+
+    fn visit_between(
+        &mut self,
+        expr: &Expr,
+        _negated: &bool,
+        low: &Expr,
+        high: &Expr,
+    ) -> Result<()> {
+        ExprTraverser::accept(expr, self)?;
+        ExprTraverser::accept(low, self)?;
+        ExprTraverser::accept(high, self)
+    }
+
+    fn visit_position(&mut self, substr_expr: &Expr, str_expr: &Expr) -> Result<()> {
+        ExprTraverser::accept(substr_expr, self)?;
+        ExprTraverser::accept(str_expr, self)
+    }
+
+    fn visit_substring(
+        &mut self,
+        expr: &Expr,
+        from: &Option<Box<Expr>>,
+        length: &Option<Box<Expr>>,
+    ) -> Result<()> {
+        ExprTraverser::accept(expr, self)?;
+
+        if let Some(from) = from {
+            ExprTraverser::accept(from, self)?;
+        }
+
+        if let Some(length) = length {
+            ExprTraverser::accept(length, self)?;
+        }
+
+        Ok(())
+    }
+
+    fn visit_extract(&mut self, _field: &DateTimeField, expr: &Expr) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_map_access(&mut self, expr: &Expr, _keys: &[Value]) -> Result<()> {
+        ExprTraverser::accept(expr, self)
+    }
+
+    fn visit_trim(
+        &mut self,
+        expr: &Expr,
+        trim_where: &Option<(TrimWhereField, Box<Expr>)>,
+    ) -> Result<()> {
+        ExprTraverser::accept(expr, self)?;
+
+        if let Some(trim_where) = trim_where {
+            ExprTraverser::accept(&trim_where.1, self)?;
+        }
+        Ok(())
+    }
+
+    fn visit_array(&mut self, exprs: &[Expr]) -> Result<()> {
+        for expr in exprs {
+            ExprTraverser::accept(expr, self)?;
+        }
+        Ok(())
+    }
+}
+
+#[async_trait]
+impl ExprVisitor for ExprRPNBuilder {
+    fn pre_visit(&mut self, expr: &Expr) -> Result<Expr> {
         Ok(expr.clone())
     }
