+++ src/common/management/src/udf/udf_mgr.rs
@@ -14,7 +14,6 @@
 
 use std::sync::Arc;
 
-use common_ast::udfs::UDFParser;
 use common_base::base::escape_for_key;
 use common_exception::ErrorCode;
 use common_exception::Result;
@@ -62,9 +61,6 @@ impl UdfApi for UdfMgr {
                 info.name.as_str()
             )));
         }
-
-        let mut udf_parser = UDFParser::default();
-        udf_parser.parse(&info.name, &info.parameters, &info.definition)?;
 
         let seq = MatchSeq::Exact(0);
         let val = Operation::Update(serde_json::to_vec(&info)?);
