+++ src/query/ast/src/lib.rs
@@ -15,6 +15,8 @@
 // TODO(xuanwo): Add crate level documents here.
 
 mod error;
+mod udf_validator;
+
 pub use error::Backtrace;
 pub use error::DisplayError;
 pub use error::Error;
@@ -33,9 +35,9 @@ pub use input::Dialect;
 pub use input::Input;
 
 mod util;
+pub use udf_validator::UDFValidator;
 pub use util::match_text;
 pub use util::match_token;
 
 pub mod ast;
 pub mod parser;
-pub mod udfs;
