--- src/query/ast/src/udfs/mod.rs
@@ -1,26 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod udf_definition;
-mod udf_expr_visitor;
-mod udf_fetcher;
-mod udf_parser;
-mod udf_transformer;
-
-pub use udf_definition::UDFDefinition;
-pub use udf_expr_visitor::UDFExprTraverser;
-pub use udf_expr_visitor::UDFExprVisitor;
-pub use udf_fetcher::UDFFetcher;
-pub use udf_parser::UDFParser;
-pub use udf_transformer::UDFTransformer;
