+++ src/query/ast/tests/it/main.rs
@@ -14,4 +14,3 @@
 
 mod parser;
 mod token;
-mod udfs;
