+++ src/query/service/src/sql/statements/query/query_schema_joined_analyzer.rs
@@ -158,8 +158,8 @@ impl JoinedSchemaAnalyzer {
 
         for table_arg in &item.args {
             table_args.push(match table_arg {
-                FunctionArg::Named { arg, .. } => analyzer.analyze_function_arg(arg)?,
-                FunctionArg::Unnamed(arg) => analyzer.analyze_function_arg(arg)?,
+                FunctionArg::Named { arg, .. } => analyzer.analyze_function_arg(arg).await?,
+                FunctionArg::Unnamed(arg) => analyzer.analyze_function_arg(arg).await?,
             });
         }
